requirements.txt
tensorflow~=2.0
fire~=0.1.3
python-binance~=0.7.1
pandas~=0.24.2
pytest~=5.0.1
pyyaml~=5.1.1

Flask~=1.1.1
Flask-Caching~=1.7.2
dash~=1.1.1
dash-daq~=0.1.0

